pnl/requirements.txt
numpy==1.26.3
pandas==2.1.4
python-dateutil==2.8.2
requests==2.31.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0